requirements.txt
matplotlib==3.5.2
onboard.client==1.8.0
pandas==1.4.2
python-dateutil==2.8.2
scikit-learn==1.0.2
python-dotenv==0.20.0